+++ actix-limitation/CHANGES.md
@@ -3,6 +3,9 @@
 ## Unreleased - 2022-xx-xx
 - Implement `Default` for `RateLimiter`.
 - `RateLimiter` is marked `#[non_exhaustive]`; use `RateLimiter::default()` instead.
+- In the middleware errors from the count function are matched and respond with `INTERNAL_SERVER_ERROR` if it's an unexpected error, instead of the default `TOO_MANY_REQUESTS`.
+- Minimum supported Rust version (MSRV) is now 1.59 due to transitive `time` dependency.
+
 
 ## 0.3.0 - 2022-07-11
 - `Limiter::builder` now takes an `impl Into<String>`.